refactor: move `config` and `utils` dirs to `/common` root dir, run migration & seeder after docker build
