+++ common/utils/db/errors.ts
