+++ src/district/district.controller.spec.ts
@@ -1,5 +1,5 @@
-import { getValues, sortArray } from '@/common/utils/array';
-import { getDistricts } from '@/common/utils/data';
+import { getValues, sortArray } from 'common/utils/array';
+import { getDistricts } from 'common/utils/data';
 import { SortOrder } from '@/sort/sort.dto';
 import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
