+++ src/district/district.service.spec.ts
@@ -1,5 +1,5 @@
-import { getDistricts } from '@/common/utils/data';
-import { getDBProviderFeatures } from '@/common/utils/db';
+import { getDistricts } from 'common/utils/data';
+import { getDBProviderFeatures } from 'common/utils/db';
 import { SortOrder } from '@/sort/sort.dto';
 import { Test, TestingModule } from '@nestjs/testing';
 import { District } from '@prisma/client';
