+++ src/island/__mocks__/island.service.ts
@@ -1,5 +1,5 @@
-import { sortArray } from '@/common/utils/array';
-import { convertCoordinate } from '@/common/utils/coordinate';
+import { sortArray } from 'common/utils/array';
+import { convertCoordinate } from 'common/utils/coordinate';
 import { Island } from '@prisma/client';
 import { IslandFindQueries } from '../island.dto';
