+++ src/main.ts
@@ -5,7 +5,7 @@ import {
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { appConfig } from '@/common/config/app';
+import { appConfig } from 'common/config/app';
 import { AppModule } from './app.module';
 
 async function bootstrap() {
