+++ src/prisma/prisma.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test } from '@nestjs/testing';
 import { PaginatorOptions, PrismaService } from './prisma.service';
 import { PrismaClient, Province } from '@prisma/client';
-import { appConfig } from '@/common/config/app';
+import { appConfig } from 'common/config/app';
 
 describe('PrismaService', () => {
   let service: PrismaService;
