+++ src/province/province.service.spec.ts
@@ -1,5 +1,5 @@
-import { getProvinces } from '@/common/utils/data';
-import { getDBProviderFeatures } from '@/common/utils/db';
+import { getProvinces } from 'common/utils/data';
+import { getDBProviderFeatures } from 'common/utils/db';
 import { PrismaService } from '@/prisma/prisma.service';
 import { SortOrder } from '@/sort/sort.dto';
 import { Test, TestingModule } from '@nestjs/testing';
