+++ src/village/village.service.ts
@@ -1,5 +1,5 @@
 import { PaginatedReturn } from '@/common/interceptor/paginate.interceptor';
-import { getDBProviderFeatures } from '@/common/utils/db';
+import { getDBProviderFeatures } from 'common/utils/db';
 import { PrismaService } from '@/prisma/prisma.service';
 import { SortService } from '@/sort/sort.service';
 import { Injectable } from '@nestjs/common';
